refactor: use lowercase hyphen mode, step 2
