+++ src/main/decorator/index.ts
@@ -1,6 +1,6 @@
-export * from './controller.bak';
-export * from './handler.bak';
-export * from './injector.bak';
-export * from './method.bak';
-export * from './parameter.bak';
-export * from './service.bak';
+export * from './controller';
+export * from './handler';
+export * from './injector';
+export * from './method';
+export * from './parameter';
+export * from './service';
