+++ src/main/handler/exception.handler.ts
@@ -1,5 +1,5 @@
-import { HTTP500Exception, HTTPException } from '../exception.temp';
-import { BaseHandler } from './BaseHandler';
+import { HTTP500Exception, HTTPException } from '../exception';
+import { BaseHandler } from './base.handler';
 
 export class ExceptionHandler extends BaseHandler {
