+++ src/main/handler/schema.handler.ts
@@ -1,4 +1,4 @@
-import { BaseHandler } from './BaseHandler';
+import { BaseHandler } from './base.handler';
 
 /**
  * Schema handler, draft.
