+++ src/main/index.ts
@@ -1,5 +1,5 @@
 export * from './@types';
 export * from './decorator';
-export * from './exception.temp';
+export * from './exception';
 export * from './handler';
-export * from './rester.temp';
+export * from './rester';
